refcator docs layout
