+++ contents/docs/integrate/client/android/index.mdx
@@ -1,5 +1,6 @@
 ---
 title: Android library
+layout: library
 sidebarTitle: Android
 sidebar: Docs
 showTitle: true
