+++ contents/docs/integrate/client/flutter/index.mdx
@@ -1,5 +1,6 @@
 ---
 title: Flutter library
+layout: library
 sidebarTitle: Flutter
 sidebar: Docs
 showTitle: true
