+++ contents/docs/integrate/client/ios/index.mdx
@@ -1,5 +1,6 @@
 ---
 title: iOS library
+layout: library
 sidebarTitle: iOS
 sidebar: Docs
 showTitle: true
