+++ contents/docs/integrate/client/js/index.mdx
@@ -1,5 +1,6 @@
 ---
 title: JavaScript library
+layout: library
 sidebarTitle: JavaScript
 sidebar: Docs
 showTitle: true
