+++ contents/docs/integrate/client/react-native/index.mdx
@@ -1,5 +1,6 @@
 ---
 title: React Native library
+layout: library
 sidebarTitle: React Native
 sidebar: Docs
 showTitle: true
