+++ contents/docs/integrate/server/elixir.mdx
@@ -1,5 +1,6 @@
 ---
 title: Elixir library
+layout: library
 sidebarTitle: Elixir
 sidebar: Docs
 showTitle: true
