+++ contents/docs/integrate/server/go.mdx
@@ -1,5 +1,6 @@
 ---
 title: Go library
+layout: library
 sidebarTitle: Go
 sidebar: Docs
 showTitle: true
