+++ contents/docs/integrate/server/java.mdx
@@ -1,5 +1,6 @@
 ---
 title: Java Library
+layout: library
 sidebarTitle: Java
 sidebar: Docs
 showTitle: true
