+++ contents/docs/integrate/server/node.mdx
@@ -1,5 +1,6 @@
 ---
 title: Node.js library
+layout: library
 sidebarTitle: Node.js
 sidebar: Docs
 showTitle: true
