+++ contents/docs/integrate/server/php.mdx
@@ -1,5 +1,6 @@
 ---
 title: PHP library
+layout: library
 sidebarTitle: PHP
 sidebar: Docs
 showTitle: true
