+++ contents/docs/integrate/server/python.mdx
@@ -1,5 +1,6 @@
 ---
 title: Python Library
+layout: library
 sidebarTitle: Python
 sidebar: Docs
 showTitle: true
