+++ contents/docs/integrate/server/ruby.mdx
@@ -1,5 +1,6 @@
 ---
 title: Ruby Library
+layout: library
 sidebarTitle: Ruby
 sidebar: Docs
 showTitle: true
