+++ src/components/Docs/Layout.tsx
@@ -1,6 +1,9 @@
-import React, { useRef } from 'react'
+import React from 'react'
 import { MDXProvider } from '@mdx-js/react'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
+
+import { animateScroll as scroll } from 'react-scroll'
+import { useLocation } from '@reach/router'
 
 import { shortcodes } from '../../mdxGlobalComponents'
 import MainSidebar from './MainSidebar'
@@ -11,15 +14,16 @@ import { Blockquote } from 'components/BlockQuote'
 import { CodeBlock } from 'components/CodeBlock'
 import CommunityQuestions from 'components/CommunityQuestions'
 import { Heading } from 'components/Heading'
-import Link from 'components/Link'
+import { push as Menu } from 'react-burger-menu'
 import { InlineCode } from 'components/InlineCode'
 import Team from 'components/Team'
 import TestimonialsTable from 'components/TestimonialsTable'
-import LibraryFeatures, { LibraryFeaturesProps } from 'components/LibraryFeatures'
-import { GitHub } from 'components/Icons/Icons'
 import { ZoomImage } from 'components/ZoomImage'
+import Layout from 'components/Layout'
+import Navigation from './Navigation'
+import Footer from './Footer'
 
-const Iframe = (props) => {
+const Iframe = (props: React.IframeHTMLAttributes<HTMLIFrameElement>) => {
     if (props.src && props.src.indexOf('youtube.com') !== -1) {
         return (
             <div style={{ position: 'relative', height: 0, paddingBottom: '56.25%' }}>
@@ -50,12 +54,14 @@ const SectionLinksTop = ({ previous, next }: SectionLinksProps) => {
 
 type DocsLayoutProps = {
     title: string
-    github?: string
-    features?: LibraryFeaturesProps['availability']
+    titleElement?: React.ReactNode
+    filePath: string
     lastUpdated: string
     menu: any
     slug: string
     body: string
+    breadcrumb: any
+    breadcrumbBase: any
     tableOfContents: any
     next: {
         name: string
@@ -67,21 +73,27 @@ type DocsLayoutProps = {
     }
     hideLastUpdated: boolean
     hideAnchor: boolean
+    contributors: any[]
+    children: React.ReactNode
 }
 
 export default function DocsLayout({
     title,
-    github,
+    titleElement,
+    filePath,
     lastUpdated,
     menu,
     slug,
     hideAnchor,
-    features,
+    breadcrumb,
+    breadcrumbBase,
     tableOfContents,
     body,
     next,
     previous,
     hideLastUpdated,
+    contributors,
+    children,
 }: DocsLayoutProps) {
     const components = {
         Team,
@@ -100,56 +112,99 @@ export default function DocsLayout({
         ...shortcodes,
     }
 
+    const { hash } = useLocation()
+
+    React.useEffect(() => {
+        if (hash) {
+            scroll.scrollMore(-50)
+        }
+    }, [])
+
     const showToc = !hideAnchor && tableOfContents?.length > 0
-    const mainEl = useRef()
+    const mainEl = React.useRef(null)
+
+    const [menuOpen, setMenuOpen] = React.useState<boolean>(false)
 
     return (
-        <div className="relative">
-            <SectionLinksTop next={next} previous={previous} />
-            <div className="dark:text-white flex max-w-screen-3xl mx-auto items-start relative z-10 mt-8">
-                <MainSidebar
-                    sticky
-                    top={90}
-                    mainEl={mainEl}
-                    menu={menu}
-                    slug={slug}
-                    className="hidden md:block w-full transition-opacity md:opacity-60 hover:opacity-100 mb-14 flex-1"
-                />
-                <article
-                    ref={mainEl}
-                    style={!showToc ? { maxWidth: '100%', paddingRight: 0 } : {}}
-                    className="w-full 2xl:max-w-[800px] xl:max-w-[650px] max-w-full pb-14 relative md:pl-16 xl:px-16 2xl:px-32 box-content overflow-auto"
-                >
-                    <section className="mb-8 relative">
-                        <h1 className="flex items-center justify-between dark:text-white text-3xl sm:text-5xl mt-0 mb-2">
-                            {title}
-                            {github && (
-                                <Link to={github}>
-                                    <GitHub className="w-8 h-8 text-black/80 hover:text-black/60 dark:text-white/80 hover:dark:text-white/60 transition-colors" />
-                                </Link>
-                            )}
-                        </h1>
-                        {!hideLastUpdated && (
-                            <p className="mt-1 mb-0 opacity-50">
-                                Last updated: <time>{lastUpdated}</time>
-                            </p>
-                        )}
-                    </section>
+        <Layout>
+            <div className="handbook-container px-4">
+                <div id="handbook-menu-wrapper">
+                    <Menu
+                        width="calc(100vw - 80px)"
+                        onClose={() => setMenuOpen(false)}
+                        customBurgerIcon={false}
+                        customCrossIcon={false}
+                        styles={{
+                            bmOverlay: {
+                                background: 'transparent',
+                            },
+                        }}
+                        pageWrapId="handbook-content-menu-wrapper"
+                        outerContainerId="handbook-menu-wrapper"
+                        overlayClassName="backdrop-blur"
+                        isOpen={menuOpen}
+                    >
+                        <MainSidebar height={'auto'} menu={menu} slug={slug} className="p-5 pb-32 md:hidden" />
+                    </Menu>
 
-                    {showToc && <MobileSidebar tableOfContents={tableOfContents} />}
+                    <Navigation
+                        title={title}
+                        filePath={filePath}
+                        next={next}
+                        previous={previous}
+                        breadcrumb={breadcrumb}
+                        breadcrumbBase={breadcrumbBase}
+                        menuOpen={menuOpen}
+                        handleMobileMenuClick={() => setMenuOpen((open) => !open)}
+                    />
+                    <div id="handbook-content-menu-wrapper">
+                        <div className="relative">
+                            <SectionLinksTop next={next} previous={previous} />
+                            <div className="dark:text-white flex max-w-screen-3xl mx-auto items-start relative z-10 mt-8">
+                                <MainSidebar
+                                    sticky
+                                    top={90}
+                                    height={undefined}
+                                    mainEl={mainEl}
+                                    menu={menu}
+                                    slug={slug}
+                                    className="hidden md:block w-full transition-opacity md:opacity-60 hover:opacity-100 mb-14 flex-1"
+                                />
+                                <article
+                                    ref={mainEl}
+                                    style={!showToc ? { maxWidth: '100%', paddingRight: 0 } : {}}
+                                    className="w-full 2xl:max-w-[800px] xl:max-w-[650px] max-w-full pb-14 relative md:pl-16 xl:px-16 2xl:px-32 box-content overflow-auto"
+                                >
+                                    <section className="mb-8 relative">
+                                        <div className="dark:text-white text-3xl sm:text-5xl mt-0">
+                                            {titleElement || <h1 className="mb-2">{title}</h1>}
+                                        </div>
+                                        {!hideLastUpdated && (
+                                            <p className="mt-1 mb-0 opacity-50">
+                                                Last updated: <time>{lastUpdated}</time>
+                                            </p>
+                                        )}
+                                    </section>
 
-                    <section className="article-content handbook-docs-content">
-                        <LibraryFeatures availability={features} />
-                        <MDXProvider components={components}>
-                            <MDXRenderer>{body}</MDXRenderer>
-                        </MDXProvider>
-                    </section>
-                    <CommunityQuestions />
-                </article>
+                                    {showToc && <MobileSidebar tableOfContents={tableOfContents} />}
 
-                {showToc && <StickySidebar top={90} tableOfContents={tableOfContents} />}
+                                    <section className="article-content handbook-docs-content">
+                                        {children}
+                                        <MDXProvider components={components}>
+                                            <MDXRenderer>{body}</MDXRenderer>
+                                        </MDXProvider>
+                                    </section>
+                                    <CommunityQuestions />
+                                </article>
+
+                                {showToc && <StickySidebar top={90} tableOfContents={tableOfContents} />}
+                            </div>
+                            {next && <SectionLinksBottom next={next} previous={previous} />}
+                        </div>
+                    </div>
+                </div>
+                <Footer title={title} filePath={filePath} contributors={contributors} />
             </div>
-            {next && <SectionLinksBottom next={next} previous={previous} />}
-        </div>
+        </Layout>
     )
 }
